# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API do Sistema Financeiro
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contas-a-receber:
    get:
      operationId: listContasAReceber
      tags: [contasAReceber]
      summary: Listar contas a receber
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, recebido, vencido, cancelado]
        - name: dataInicio
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: dataFim
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: busca
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Lista de contas a receber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAReceberListResponse"
    post:
      operationId: createContaAReceber
      tags: [contasAReceber]
      summary: Criar conta a receber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContaAReceberInput"
      responses:
        "201":
          description: Conta a receber criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAReceber"
  /contas-a-receber/{id}:
    get:
      operationId: getContaAReceber
      tags: [contasAReceber]
      summary: Buscar conta a receber por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conta a receber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAReceber"
        "404":
          description: Não encontrado
    patch:
      operationId: updateContaAReceber
      tags: [contasAReceber]
      summary: Atualizar conta a receber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContaAReceberUpdate"
      responses:
        "200":
          description: Conta a receber atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAReceber"
    delete:
      operationId: deleteContaAReceber
      tags: [contasAReceber]
      summary: Excluir conta a receber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Excluído
  /contas-a-receber/{id}/receber:
    post:
      operationId: receberContaAReceber
      tags: [contasAReceber]
      summary: Marcar conta a receber como recebida
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecebimentoInput"
      responses:
        "200":
          description: Conta recebida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAReceber"

  # ─────────────────────────── FLUXO DE CAIXA ───────────────────────────
...
components:
  schemas:
...
    ContaAReceber:
      type: object
      required: [id, descricao, valor, dataVencimento, status, createdAt]
      properties:
        id:
          type: integer
        descricao:
          type: string
        cliente:
          type: ["string", "null"]
        valor:
          type: number
        dataVencimento:
          type: string
          format: date
        dataRecebimento:
          type: ["string", "null"]
          format: date
        valorRecebido:
          type: ["number", "null"]
        status:
          type: string
          enum: [pendente, recebido, vencido, cancelado]
        observacoes:
          type: ["string", "null"]
        categoriaId:
          type: ["integer", "null"]
        categoriaNome:
          type: ["string", "null"]
        contaId:
          type: ["integer", "null"]
        contaNome:
          type: ["string", "null"]
        recorrente:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ContaAReceberInput:
      type: object
      required: [descricao, valor, dataVencimento]
      properties:
        descricao:
          type: string
          minLength: 1
        cliente:
          type: string
        valor:
          type: number
          minimum: 0.01
        dataVencimento:
          type: string
          format: date
        observacoes:
          type: string
        categoriaId:
          type: ["integer", "null"]
        contaId:
          type: ["integer", "null"]
        recorrente:
          type: boolean
    ContaAReceberUpdate:
      type: object
      properties:
        descricao:
          type: string
        cliente:
          type: string
        valor:
          type: number
        dataVencimento:
          type: string
          format: date
        status:
          type: string
          enum: [pendente, recebido, vencido, cancelado]
        observacoes:
          type: string
        categoriaId:
          type: ["integer", "null"]
        contaId:
          type: ["integer", "null"]
        recorrente:
          type: boolean
    RecebimentoInput:
      type: object
      required: [dataRecebimento]
      properties:
        dataRecebimento:
          type: string
          format: date
        valorRecebido:
          type: number
        contaId:
          type: ["integer", "null"]
    ContaAReceberListResponse:
      type: object
      required: [data, total, page, limit, totalPendente, totalVencido]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ContaAReceber"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPendente:
          type: number
        totalVencido:
          type: number

    # ─────── FLUXO DE CAIXA ───────
...