# lib/api-zod/src/generated/types/contaAReceber.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API do Sistema Financeiro
 * OpenAPI spec version: 0.1.0
 */
import type { ContaAReceberStatus } from './contaAReceberStatus';

export interface ContaAReceber {
  id: number;
  descricao: string;
  /** @nullable */
  cliente?: string | null;
  valor: number;
  dataVencimento: Date;
  /** @nullable */
  dataRecebimento?: Date | null;
  /** @nullable */
  valorRecebido?: number | null;
  status: ContaAReceberStatus;
  /** @nullable */
  observacoes?: string | null;
  /** @nullable */
  categoriaId?: number | null;
  /** @nullable */
  categoriaNome?: string | null;
  /** @nullable */
  contaId?: number | null;
  /** @nullable */
  contaNome?: string | null;
  recorrente?: boolean;
  createdAt: Date;
}