# lib/api-zod/src/generated/types/contaAReceberUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API do Sistema Financeiro
 * OpenAPI spec version: 0.1.0
 */
import type { ContaAReceberUpdateStatus } from './contaAReceberUpdateStatus';

export interface ContaAReceberUpdate {
  descricao?: string;
  cliente?: string;
  valor?: number;
  dataVencimento?: Date;
  status?: ContaAReceberUpdateStatus;
  observacoes?: string;
  /** @nullable */
  categoriaId?: number | null;
  /** @nullable */
  contaId?: number | null;
  recorrente?: boolean;
}